requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alpha_vantage==3.0.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
eodhd==1.0.30
et-xmlfile==1.1.0
-e git+https://github.com/mattsmith321/fin-ds.git@b7bc93f922f5a4c57f6d636be7caec508b33a34f#egg=fin_ds
fonttools==4.48.1
frozendict==2.4.0
frozenlist==1.4.1
html5lib==1.1
idna==3.6
inflection==0.5.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
multitasking==0.0.11
Nasdaq-Data-Link==1.0.4
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.3
peewee==3.17.1
pillow==10.2.0
platformdirs==4.3.6
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
rich==13.7.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
tiingo==0.15.6
tzdata==2023.4
urllib3==2.2.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
yfinance==0.2.43
